fix crashes due to char -> unsigned int sign extension
	* undo.c, redisplay.c, mac_windows.m, glyphs.c, editcommands.c:
	fix some sign-extension-and-overflow issues introduced when
	eliminating u_char, and eliminate the remaining uses of u_char

+++ src/editcommands.c
@@ -336,7 +336,7 @@ unchanged.
 	    str = VBUFFER(tx)->lines[linenum].ln_Line + col;
 	    while(col++ < llen)
 	    {
-		unsigned int c = *str;
+		uint8_t c = *str;
 		*str++ = (c < tablen) ? rep_STR(table)[c] : c;
 	    }
 	    linenum++;
@@ -345,7 +345,7 @@ unchanged.
 	str = VBUFFER(tx)->lines[linenum].ln_Line + col;
 	while(col++ < VCOL(end))
 	{
-	    unsigned int c = *str;
+	    uint8_t c = *str;
 	    *str++ = (c < tablen) ? rep_STR(table)[c] : c;
 	}
 	return(Qt);
